// src/lib/types.ts
export type MagicCard = {
  name: string
  type: string
  manaCost: string
  cmc: number
  setName: string
  rarity: string
  imageUrl: string
  text: string
  set: string
  artist: string
  number: string
  id: string
  multiverseid: string
  flavor: string
  colors: string[]
  types: string[]
  subtypes: string[]
  supertypes: string[]
  printings: string[]
  power: string
  toughness: string
}